refactor share listeners into lib
