+++ src/libs/Share/index.android.js
@@ -1,0 +1,11 @@
+import ShareMenu from 'react-native-share-menu';
+import navigateToShare from './navigateToShare';
+
+const registerListener = () => {
+    ShareMenu.getInitialShare(navigateToShare);
+    return ShareMenu.addNewShareListener(navigateToShare);
+};
+
+export default {
+    registerListener,
+};
