+++ src/libs/Share/index.js
@@ -1,0 +1,3 @@
+export default {
+    registerListener: () => ({remove: () => {}}),
+};
